+++ packages/framework/src/Form/Admin/Transport/TransportFormType.php
@@ -4,7 +4,6 @@ namespace Shopsys\FrameworkBundle\Form\Admin\Transport;
 
 use Ivory\CKEditorBundle\Form\Type\CKEditorType;
 use Shopsys\FormTypesBundle\YesNoType;
-use Shopsys\FrameworkBundle\Component\Transformers\IndexedBooleansToArrayOfIndexesTransformer;
 use Shopsys\FrameworkBundle\Form\DisplayOnlyType;
 use Shopsys\FrameworkBundle\Form\DomainsType;
 use Shopsys\FrameworkBundle\Form\GroupType;
@@ -32,8 +31,10 @@ class TransportFormType extends AbstractType
      */
     private $paymentFacade;
 
-    public function __construct(VatFacade $vatFacade, PaymentFacade $paymentFacade)
-    {
+    public function __construct(
+        VatFacade $vatFacade,
+        PaymentFacade $paymentFacade
+    ) {
         $this->vatFacade = $vatFacade;
         $this->paymentFacade = $paymentFacade;
     }
@@ -70,7 +71,7 @@ class TransportFormType extends AbstractType
                 ],
                 'label' => t('Name'),
             ])
-            ->add('domains', DomainsType::class, [
+            ->add('enabled', DomainsType::class, [
                 'required' => false,
                 'label' => t('Display on'),
             ])
@@ -98,8 +99,6 @@ class TransportFormType extends AbstractType
                 'empty_message' => t('You have to create some payment first.'),
                 'label' => t('Available payment methods'),
             ]);
-        $builderBasicInformationGroup->get('domains')
-            ->addModelTransformer(new IndexedBooleansToArrayOfIndexesTransformer());
 
         $builderAdditionalInformationGroup = $builder->create('additionalInformation', GroupType::class, [
             'label' => t('Additional information'),